Split helloworld -> numeric and bytestring version
The new executable `plutus-helloworld-bytestring` succeeds in doing what
the previous `plutus-helloworld` attempted but (apparently) failed to
do: implement a validator that would only accept the bytestring datum
"Hello World!".

It does this by using a parametric typed validator with an extra argument
`keyword` to which the datum is compared. During compilation of the
validator, the bytestring keyword "Hello World!" injected via
`applyCode` and `liftCode`.

Another new executable `plutus-helloworld-numeric` simplifies the
validator significantly by making the keyword numeric (123), which
avoids the need for a parametric typed validator. This is possible
because integer values of the Plutus `Data` type can be constructed
directly, unlike bytestring values of this type (I think).